Using braces for multiline if statements.
